+++ src/app/app.module.ts
@@ -4,10 +4,14 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { FormsModule } from '@angular/forms';
+import { HomeComponent } from './landing/componentes/home/home.component';
+import { BotonDePagoComponent } from './landing/componentes/boton-de-pago/boton-de-pago.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    HomeComponent,
+    BotonDePagoComponent,
   ],
   imports: [
     BrowserModule,
