+++ src/app/landing/componentes/home/home.component.html
@@ -1,0 +1,1 @@
+<button class="btn btn-success" (click)="payMethod()" >Donación</button>
